# database/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "黑马头条自媒体人管理系统 API",
    "description": "黑马头条是一款面向自媒体创作者的内容管理与数据分析平台。核心功能包括：登录系统、图文数据统计、发布文章、内容列表管理、素材管理、粉丝管理。所有枚举字段均使用 TINYINT 数值存储，与数据库保持一致。",
    "version": "1.0.0",
    "contact": {
      "name": "开发团队"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "本地开发服务器"
    }
  ],
  "security": [
    { "bearerAuth": [] }
  ],
...
  "paths": {
    "/auth/login": {
      "post": {
        "tags": ["认证"],
        "summary": "用户登录",
        "description": "使用用户名和密码登录，返回 JWT Token。需勾选用户协议。",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "登录成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LoginResponse" }
              }
            }
          },
          "401": { "description": "用户名或密码错误" },
          "422": { "description": "未同意用户协议" }
        }
      }
    },
    "/auth/register": {
      "post": {
        "tags": ["认证"],
        "summary": "用户注册",
        "description": "注册新的自媒体创作者账号。",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "注册成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserResponse" }
              }
            }
          },
          "409": { "description": "用户名已存在" }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "tags": ["认证"],
        "summary": "刷新Token",
        "description": "使用当前有效的 Token 换取新 Token。",
        "responses": {
          "200": {
            "description": "刷新成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TokenResponse" }
              }
            }
          },
          "401": { "description": "Token 已过期" }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": ["认证"],
        "summary": "获取当前用户信息",
        "description": "根据 JWT Token 获取当前登录用户的信息。",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserDetailResponse" }
              }
            }
          },
          "401": { "description": "未认证或 Token 失效" }
        }
      }
    },
...
    "/articles": {
      "get": {
        "tags": ["文章管理"],
        "summary": "文章列表",
        "description": "分页查询当前用户的文章列表，支持按状态、频道、关键字、日期范围筛选。默认过滤已软删除的文章。",
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 }, "description": "页码" },
          { "name": "pageSize", "in": "query", "schema": { "type": "integer", "default": 10 }, "description": "每页条数" },
          { "name": "status", "in": "query", "schema": { "type": "integer", "enum": [0, 1, 2, 3, 4, 5] }, "description": "文章状态筛选" },
          { "name": "channelId", "in": "query", "schema": { "type": "integer", "format": "int64" }, "description": "频道ID筛选" },
          { "name": "keyword", "in": "query", "schema": { "type": "string" }, "description": "标题模糊搜索" },
          { "name": "startDate", "in": "query", "schema": { "type": "string", "format": "date" }, "description": "创建时间起始" },
          { "name": "endDate", "in": "query", "schema": { "type": "string", "format": "date" }, "description": "创建时间截止" },
          { "name": "sortBy", "in": "query", "schema": { "type": "string", "default": "createdAt" }, "description": "排序字段" },
          { "name": "order", "in": "query", "schema": { "type": "string", "enum": ["asc", "desc"], "default": "desc" }, "description": "排序方向" }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ArticleListResponse" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["文章管理"],
        "summary": "创建文章",
        "description": "创建文章。status=0 为保存草稿，status=1 为提交审核。",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ArticleCreateRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "创建成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ArticleDetailResponse" }
              }
            }
          }
        }
      }
    },
    "/articles/enums/{field}": {
      "get": {
        "tags": ["文章管理"],
        "summary": "获取枚举值",
        "description": "获取文章状态或封面类型的枚举值列表。field 可选值: status / coverType。",
        "parameters": [
          {
            "name": "field",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "enum": ["status", "coverType"] },
            "description": "枚举字段名: status=文章状态, coverType=封面类型"
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/EnumResponse" }
              }
            }
          }
        }
      }
    },
    "/articles/{id}": {
      "get": {
        "tags": ["文章管理"],
        "summary": "文章详情",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ArticleDetailResponse" }
              }
            }
          },
          "404": { "description": "文章不存在或已删除" }
        }
      },
      "put": {
        "tags": ["文章管理"],
        "summary": "更新文章",
        "description": "更新文章内容。仅草稿(0)和审核失败(3)状态可编辑。",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ArticleUpdateRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ArticleDetailResponse" }
              }
            }
          },
          "422": { "description": "当前状态不允许编辑" }
        }
      },
      "delete": {
        "tags": ["文章管理"],
        "summary": "删除文章（软删除）",
        "description": "仅草稿(0)和审核失败(3)状态可删除。软删除后列表查询自动过滤。",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }
        ],
        "responses": {
          "200": { "description": "删除成功" },
          "422": { "description": "当前状态不允许删除" }
        }
      }
    },
    "/articles/{id}/status": {
      "patch": {
        "tags": ["文章管理"],
        "summary": "变更文章状态",
        "description": "执行文章状态流转：提交审核、审核通过/失败、上架、下架等。",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ArticleStatusChangeRequest" }
            }
          }
        },
        "responses": {
          "200": { "description": "状态变更成功" },
          "422": { "description": "不允许的状态流转" }
        }
      }
    },
    "/articles/{id}/restore": {
      "patch": {
        "tags": ["文章管理"],
        "summary": "恢复文章",
        "description": "恢复已软删除的文章。",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }
        ],
        "responses": {
          "200": { "description": "恢复成功" },
          "404": { "description": "文章不存在或未被删除" }
        }
      }
    },
    "/articles/{id}/covers": {
      "get": {
        "tags": ["文章管理"],
        "summary": "获取文章封面",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ArticleCoverResponse" }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["文章管理"],
        "summary": "设置文章封面",
        "description": "设置或更新文章封面图片，替换全部现有封面。coverType=0(单图)时传1个, =1(三图)时传3个, =2(无图)时空数组。",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ArticleCoverSetRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "设置成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ArticleCoverResponse" }
              }
            }
          },
          "422": { "description": "封面数量与coverType不匹配" }
        }
      }
    },
...
    "/article-stats/overview": {
      "get": {
        "tags": ["文章统计"],
        "summary": "统计概览卡片",
        "description": "获取图文数据概览4项统计卡片：图文发布量、点赞数量、收藏数量、转发数量。",
        "parameters": [
          { "name": "startDate", "in": "query", "schema": { "type": "string", "format": "date" }, "description": "起始日期" },
          { "name": "endDate", "in": "query", "schema": { "type": "string", "format": "date" }, "description": "截止日期" }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ArticleStatsOverviewResponse" }
              }
            }
          }
        }
      }
    },
    "/article-stats": {
      "get": {
        "tags": ["文章统计"],
        "summary": "文章统计列表",
        "description": "分页查询当前用户所有文章的每日统计数据。",
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "pageSize", "in": "query", "schema": { "type": "integer", "default": 10 } },
          { "name": "startDate", "in": "query", "schema": { "type": "string", "format": "date" } },
          { "name": "endDate", "in": "query", "schema": { "type": "string", "format": "date" } },
          { "name": "sortBy", "in": "query", "schema": { "type": "string", "default": "readCount" } },
          { "name": "order", "in": "query", "schema": { "type": "string", "enum": ["asc", "desc"], "default": "desc" } }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ArticleStatsListResponse" }
              }
            }
          }
        }
      }
    },
    "/article-stats/{articleId}": {
      "get": {
        "tags": ["文章统计"],
        "summary": "文章详情统计",
        "description": "获取单篇文章的详细统计数据，含8项指标卡片和每日明细。",
        "parameters": [
          { "name": "articleId", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } },
          { "name": "startDate", "in": "query", "schema": { "type": "string", "format": "date" } },
          { "name": "endDate", "in": "query", "schema": { "type": "string", "format": "date" } }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ArticleStatsDetailResponse" }
              }
            }
          }
        }
      }
    },
    "/article-stats/{articleId}/read-sources": {
      "get": {
        "tags": ["文章统计"],
        "summary": "阅读来源分析",
        "description": "获取文章的阅读来源分布，对应前端环形图。",
        "parameters": [
          { "name": "articleId", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } },
          { "name": "statDate", "in": "query", "schema": { "type": "string", "format": "date" }, "description": "统计日期, 默认最新一天" }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ReadSourcesResponse" }
              }
            }
          }
        }
      }
    },
    "/article-stats/{articleId}/read-completion": {
      "get": {
        "tags": ["文章统计"],
        "summary": "阅读完成度分析",
        "description": "获取文章的阅读完成度分布，对应前端环形图。",
        "parameters": [
          { "name": "articleId", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } },
          { "name": "statDate", "in": "query", "schema": { "type": "string", "format": "date" }, "description": "统计日期, 默认最新一天" }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ReadCompletionResponse" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "用户登录后获取的 JWT Token"
      }
    },
    "schemas": {
      "ApiResponse": {
        "type": "object",
        "properties": {
          "code": { "type": "integer", "example": 200 },
          "message": { "type": "string", "example": "success" }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["username", "password", "agreeTerms"],
        "properties": {
          "username": { "type": "string", "description": "用户名", "example": "zhangsan", "maxLength": 64 },
          "password": { "type": "string", "description": "登录密码", "example": "123456" },
          "agreeTerms": { "type": "boolean", "description": "是否同意用户协议和隐私政策", "example": true }
        }
      },
      "RegisterRequest": {
        "type": "object",
        "required": ["username", "password", "agreeTerms"],
        "properties": {
          "username": { "type": "string", "maxLength": 64, "description": "用户名(唯一)", "example": "zhangsan" },
          "password": { "type": "string", "description": "登录密码", "example": "123456" },
          "agreeTerms": { "type": "boolean", "description": "是否同意用户协议和隐私政策", "example": true }
        }
      },
      "UserResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/ApiResponse" },
          {
            "type": "object",
            "properties": {
              "data": { "$ref": "#/components/schemas/User" }
            }
          }
        ]
      },
      "UserDetailResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/ApiResponse" },
          {
            "type": "object",
            "properties": {
              "data": { "$ref": "#/components/schemas/UserDetail" }
            }
          }
        ]
      },
      "LoginResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/ApiResponse" },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "object",
                "properties": {
                  "token": { "type": "string", "description": "JWT Token" },
                  "expiresAt": { "type": "string", "format": "date-time", "description": "过期时间" },
                  "user": { "$ref": "#/components/schemas/User" }
                }
              }
            }
          }
        ]
      },
      "TokenResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/ApiResponse" },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "object",
                "properties": {
                  "token": { "type": "string" },
                  "expiresAt": { "type": "string", "format": "date-time" }
                }
              }
            }
          }
        ]
      },
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64", "description": "用户ID" },
          "username": { "type": "string", "description": "用户名" },
          "avatar": { "type": "string", "nullable": true, "description": "头像URL" },
          "status": { "type": "integer", "enum": [0, 1], "description": "账号状态: 1=正常 0=禁用" },
          "lastLoginAt": { "type": "string", "format": "date-time", "nullable": true, "description": "最后登录时间" },
          "createdAt": { "type": "string", "format": "date-time", "description": "创建时间" }
        }
      },
      "UserDetail": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "username": { "type": "string" },
          "avatar": { "type": "string", "nullable": true },
          "status": { "type": "integer", "enum": [0, 1], "description": "1=正常 0=禁用" },
          "lastLoginAt": { "type": "string", "format": "date-time", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "ArticleCoverImage": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "materialId": { "type": "integer", "format": "int64", "description": "素材ID" },
          "url": { "type": "string", "description": "封面图片URL", "example": "https://cdn.example.com/materials/2.jpg" },
          "sortOrder": { "type": "integer", "description": "排序(单图=0, 三图依次=0/1/2)" }
        }
      },
      "Article": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "userId": { "type": "integer", "format": "int64" },
          "channelId": { "type": "integer", "format": "int64", "nullable": true },
          "channelName": { "type": "string", "nullable": true, "description": "频道名称(关联查询)" },
          "title": { "type": "string", "maxLength": 255 },
          "content": { "type": "string", "nullable": true, "description": "文章正文(JSON字符串格式)" },
          "tag": { "type": "string", "maxLength": 20, "nullable": true },
          "coverType": { "type": "integer", "enum": [0, 1, 2], "description": "0=单图 1=三图 2=无图" },
          "status": { "type": "integer", "enum": [0, 1, 2, 3, 4, 5], "description": "0=草稿 1=待审核 2=审核通过 3=审核失败 4=已上架 5=已下架" },
          "reviewComment": { "type": "string", "maxLength": 500, "nullable": true },
          "scheduledAt": { "type": "string", "format": "date-time", "nullable": true },
          "publishedAt": { "type": "string", "format": "date-time", "nullable": true },
          "deletedAt": { "type": "string", "format": "date-time", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "coverImages": { "type": "array", "items": { "$ref": "#/components/schemas/ArticleCoverImage" } }
        }
      },
      "ArticleListArticle": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "title": { "type": "string" },
          "channelId": { "type": "integer", "format": "int64", "nullable": true },
          "channelName": { "type": "string", "nullable": true },
          "tag": { "type": "string", "nullable": true },
          "coverType": { "type": "integer", "enum": [0, 1, 2] },
          "status": { "type": "integer", "enum": [0, 1, 2, 3, 4, 5] },
          "reviewComment": { "type": "string", "nullable": true },
          "scheduledAt": { "type": "string", "format": "date-time", "nullable": true },
          "publishedAt": { "type": "string", "format": "date-time", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "coverImages": { "type": "array", "items": { "$ref": "#/components/schemas/ArticleCoverImage" } },
          "stats": { "$ref": "#/components/schemas/ArticleStatsBrief" }
        }
      },
...
      "ArticleDetailResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/ApiResponse" },
          { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/Article" } } }
        ]
      },
      "ArticleListResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/ApiResponse" },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "object",
                "properties": {
                  "list": { "type": "array", "items": { "$ref": "#/components/schemas/ArticleListArticle" } },
                  "total": { "type": "integer" },
                  "page": { "type": "integer" },
                  "pageSize": { "type": "integer" }
                }
              }
            }
          }
        ]
      },
      "ArticleCreateRequest": {
        "type": "object",
        "required": ["title", "coverType"],
        "properties": {
          "channelId": { "type": "integer", "format": "int64", "nullable": true, "description": "频道ID" },
          "title": { "type": "string", "maxLength": 255, "description": "文章标题" },
          "content": { "type": "string", "nullable": true, "description": "文章正文(JSON字符串格式, 包含富文本格式和图片链接)" },
          "tag": { "type": "string", "maxLength": 20, "nullable": true, "description": "自定义标签" },
          "coverType": { "type": "integer", "enum": [0, 1, 2], "description": "0=单图 1=三图 2=无图" },
          "status": { "type": "integer", "enum": [0, 1], "default": 0, "description": "0=草稿 1=提交审核" },
          "scheduledAt": { "type": "string", "format": "date-time", "nullable": true, "description": "定时发布时间" },
          "coverMaterialIds": { "type": "array", "items": { "type": "integer", "format": "int64" }, "description": "封面素材ID数组(单图1个,三图3个,无图空数组)" }
        }
      },
      "ArticleUpdateRequest": {
        "type": "object",
        "properties": {
          "channelId": { "type": "integer", "format": "int64", "nullable": true },
          "title": { "type": "string", "maxLength": 255 },
          "content": { "type": "string", "nullable": true },
          "tag": { "type": "string", "maxLength": 20, "nullable": true },
          "coverType": { "type": "integer", "enum": [0, 1, 2] }
        }
      },
      "ArticleStatusChangeRequest": {
        "type": "object",
        "required": ["status"],
        "properties": {
          "status": { "type": "integer", "enum": [0, 1, 2, 3, 4, 5], "description": "目标状态" },
          "comment": { "type": "string", "maxLength": 500, "nullable": true, "description": "审核意见(驳回时填写)" }
        }
      },
      "ArticleCoverResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/ApiResponse" },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "object",
                "properties": {
                  "coverType": { "type": "integer", "enum": [0, 1, 2] },
                  "images": { "type": "array", "items": { "$ref": "#/components/schemas/ArticleCoverImage" } }
                }
              }
            }
          }
        ]
      },
      "ArticleCoverSetRequest": {
        "type": "object",
        "required": ["coverType"],
        "properties": {
          "coverType": { "type": "integer", "enum": [0, 1, 2], "description": "0=单图 1=三图 2=无图" },
          "materialIds": { "type": "array", "items": { "type": "integer", "format": "int64" }, "description": "素材ID数组(coverType=0时1个,=1时3个,=2时空数组)" }
        }
      },
...
      "ArticleStatsOverviewResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/ApiResponse" },
          { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/ArticleStatsOverview" } } }
        ]
      },
...
      "ArticleStatsListResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/ApiResponse" },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "object",
                "properties": {
                  "list": { "type": "array", "items": { "$ref": "#/components/schemas/ArticleStatsItem" } },
                  "total": { "type": "integer" },
                  "page": { "type": "integer" },
                  "pageSize": { "type": "integer" }
                }
              }
            }
          }
        ]
      },
...
      "ArticleStatsDetailResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/ApiResponse" },
          { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/ArticleStatsDetail" } } }
        ]
      },
...
      "ReadSourcesResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/ApiResponse" },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "object",
                "properties": {
                  "articleId": { "type": "integer", "format": "int64" },
                  "statDate": { "type": "string", "format": "date" },
                  "sources": { "type": "array", "items": { "$ref": "#/components/schemas/ReadSource" } }
                }
              }
            }
          }
        ]
      },
...
      "ReadCompletionResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/ApiResponse" },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "object",
                "properties": {
                  "articleId": { "type": "integer", "format": "int64" },
                  "statDate": { "type": "string", "format": "date" },
                  "completions": { "type": "array", "items": { "$ref": "#/components/schemas/ReadCompletion" } }
                }
              }
            }
          }
        ]
      },
...
      "EnumValue": {
        "type": "object",
        "properties": {
          "value": { "type": "integer" },
          "label": { "type": "string" },
          "color": { "type": "string", "nullable": true }
        }
      },
      "EnumResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/ApiResponse" },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "object",
                "properties": {
                  "field": { "type": "string" },
                  "values": { "type": "array", "items": { "$ref": "#/components/schemas/EnumValue" } }
                }
              }
            }
          }
        ]
      }
    }
  }
}